Provisioning connected with DocDB
